clean up class length issues
